Split out image and transcription too
